requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==3.14.1
djangorestframework-simplejwt==5.2.2
